requirements.txt
flask
flask-cors
firebase-admin
joblib
scikit-learn
pandas
numpy
gunicorn
python-dotenv
torch
torchvision
efficientnet-pytorch
pillow